Integrate Werners & Jans double branch
Move from float to double
Further suggestions for float -> double move
Moved Tools2D from float to double
More suggestions for float->double move from Gui subdirectory
Changes to FEM constraint visuals for float->double move
Suggested changes for float -> double move
Suggestions for Part module moving float -> double

+++ src/App/MeasureDistance.cpp
@@ -35,9 +35,9 @@ PROPERTY_SOURCE(App::MeasureDistance, App::DocumentObject)
 
 MeasureDistance::MeasureDistance() 
 {
-    ADD_PROPERTY_TYPE(P1,(Base::Vector3f()),"Measurement",Prop_None,"First point of measurement");
-    ADD_PROPERTY_TYPE(P2,(Base::Vector3f()),"Measurement",Prop_None,"Second point of measurement");
-    ADD_PROPERTY_TYPE(Distance,(0.0f)      ,"Measurement",App::PropertyType(Prop_ReadOnly|Prop_Output),
+    ADD_PROPERTY_TYPE(P1,(Base::Vector3d()),"Measurement",Prop_None,"First point of measurement");
+    ADD_PROPERTY_TYPE(P2,(Base::Vector3d()),"Measurement",Prop_None,"Second point of measurement");
+    ADD_PROPERTY_TYPE(Distance,(0.0)       ,"Measurement",App::PropertyType(Prop_ReadOnly|Prop_Output),
                                             "Distance between the points");
 
 }
