+++ src/Mod/Fem/Gui/ViewProviderFemConstraintBearing.cpp
@@ -118,9 +118,9 @@ void ViewProviderFemConstraintBearing::updateData(const App::Property* prop)
         pShapeSep->removeAllChildren();
 
         // This should always point outside of the cylinder
-        Base::Vector3f normal = pcConstraint->NormalDirection.getValue();        
-        Base::Vector3f base = pcConstraint->BasePoint.getValue();
-        float radius = pcConstraint->Radius.getValue();
+        Base::Vector3d normal = pcConstraint->NormalDirection.getValue();        
+        Base::Vector3d base = pcConstraint->BasePoint.getValue();
+        double radius = pcConstraint->Radius.getValue();
         base = base + radius * normal;
 
         SbVec3f b(base.x, base.y, base.z);
@@ -132,9 +132,9 @@ void ViewProviderFemConstraintBearing::updateData(const App::Property* prop)
     } else if (strcmp(prop->getName(),"AxialFree") == 0) {
         if (pShapeSep->getNumChildren() > 0) {
             // Change the symbol
-            Base::Vector3f normal = pcConstraint->NormalDirection.getValue();
-            Base::Vector3f base = pcConstraint->BasePoint.getValue();
-            float radius = pcConstraint->Radius.getValue();
+            Base::Vector3d normal = pcConstraint->NormalDirection.getValue();
+            Base::Vector3d base = pcConstraint->BasePoint.getValue();
+            double radius = pcConstraint->Radius.getValue();
             base = base + radius * normal;
 
             SbVec3f b(base.x, base.y, base.z);
